+++ examples/fc_networks.py
@@ -106,7 +106,7 @@ if scope_name and 300 <= oneview_client.api_version <= 500:
     try:
         fc_with_scope = fc_network.patch('replace',
                                          '/scopeUris',
-                                         [scope['uri']])
+                                         [scope.data['uri']])
         pprint(fc_with_scope)
     except HPOneViewException, e:
         print(e)